requirements.txt
argparse
easydict
h5py
matplotlib
numpy
open3d #==0.9 # if you're on newer versions of python / mamba / conda, installing may throw an error. installing latest open3d seems to work
opencv-python
pyyaml
scipy
tensorboardX
timm==0.4.5  # if you're on newer versions of python / mamba / conda, installing may throw an error. installing latest timm seems to work 
tqdm
transforms3d
einops
torchist
